update tests & docstrings
